added bgcd pre base
still missing the code for updating the state files with the cd data;
fixed flag parsing error.

for the bgcd, the way that the updating works is that every n seconds,
the changes are made, this is to prevent a lot of unnecessary resources
from being used. once all the possible updates have been made (without
getting over the desired time), a new system will begin which is very
oftenly checking if the time is over, additional data will be updated
after.