changed the flag system, added -n for bgcd
the new way in which the flags are processed is simply to make the code
a little bit cleaner when it gets large, it basically consists of first
defining the mode in a very robust way, removing that arg from the args
vector and then calling that mode's own function for processing its own
flags.

the -n flag in bgcd allows to specify the name of that cd, because i did
not want it to make the cli confusing by something like 'ruti cd -n 1
2', in this example the name would be 1 and the len would be 2, i did
not want things like this to happen, so it takes as a name something
that does not start in a number. it also only takes in consideration the
very first possible name, this means that in 'ruti cd 1 -n hi bye', the
name would be: he